+++ pyqrcode/__init__.py
@@ -40,9 +40,9 @@ Examples:
         >>> number.png('big-number.png')
 """
 from __future__ import absolute_import, division, print_function, with_statement, unicode_literals
-
 import pyqrcode.tables
 import pyqrcode.builder as builder
+from pyqrcode import writers
 
 try:
     str = unicode  # Python 2
@@ -420,7 +420,7 @@ class QRCode:
             >>> print(code.get_png_size(5))
             155
         """
-        return builder._get_png_size(self.version, scale, quiet_zone)
+        return writers._get_png_size(self.version, scale, quiet_zone)
 
     def png(self, file, scale=1, module_color=(0, 0, 0, 255),
             background=(255, 255, 255, 255), quiet_zone=4):
@@ -468,7 +468,7 @@ class QRCode:
                          module_color=(0x66, 0x33, 0x0),      #Dark brown
                          background=(0xff, 0xff, 0xff, 0x88)) #50% transparent white
         """
-        builder._png(self.code, self.version, file, scale,
+        writers.png(self.code, self.version, file, scale,
                      module_color, background, quiet_zone)
 
     def png_as_base64_str(self, scale=1, module_color=(0, 0, 0, 255),
@@ -535,7 +535,7 @@ class QRCode:
         left settable because such a wide quiet zone is unnecessary in many
         applications where the QR code is not being printed.
         """
-        return builder._xbm(self.code, scale, quiet_zone)
+        return writers.xbm(self.code, scale, quiet_zone)
 
     def svg(self, file, scale=1, module_color='#000', background=None,
             quiet_zone=4, xmldecl=True, svgns=True, title=None,
@@ -591,7 +591,7 @@ class QRCode:
             >>> code.svg('live-organ-transplants.svg', scale=4,
                          module_color='brown', background='0xFFFFFF')
         """
-        builder._svg(self.code, self.version, file, scale=scale, 
+        writers.svg(self.code, self.version, file, scale=scale,
                      module_color=module_color, background=background,
                      quiet_zone=quiet_zone, xmldecl=xmldecl, svgns=svgns, 
                      title=title, svgclass=svgclass, lineclass=lineclass,
@@ -629,7 +629,7 @@ class QRCode:
             >>> out = io.StringIO()
             >>> qr.eps(out, module_color=(.4, .4, .4))
         """
-        builder._eps(self.code, self.version, file, scale, module_color,
+        writers.eps(self.code, self.version, file, scale, module_color,
                      background, quiet_zone)
 
     def terminal(self, module_color='default', background='reverse',
@@ -677,7 +677,7 @@ class QRCode:
             >>> text = code.terminal()
             >>> print(text)
         """
-        return builder._terminal(self.code, module_color, background,
+        return writers.terminal(self.code, module_color, background,
                                  quiet_zone)
 
     def text(self, quiet_zone=4):
@@ -696,5 +696,4 @@ class QRCode:
             >>> text = code.text()
             >>> print(text)
         """
-        return builder._text(self.code, quiet_zone)
-
+        return writers.text(self.code, quiet_zone)
